+++ tests/testthat/test-cusum.R
@@ -9,15 +9,15 @@ test_that("cusum with no cases in training period is skipped", {
 
   # this groups has cases
   dat_0_4 <- dat_agg %>% dplyr::filter(age_group == "00-04")
-  expect_no_error(get_signals_cusum(dat_0_4, number_of_weeks = 21))
+  expect_no_error(get_signals_cusum(dat_0_4, number_of_time_units = 21))
 
   # this groups has no cases between 2023-W2 to 2023-W8
   dat_15_19 <- dat_agg %>% dplyr::filter(age_group == "15-19")
-  expect_null(get_signals_cusum(dat_15_19, number_of_weeks = 21))
+  expect_null(get_signals_cusum(dat_15_19, number_of_time_units = 21))
 
   # warning from signal wrapper
   expect_warning(
-    get_signals_all(dat, "cusum", stratification = "age_group", number_of_weeks = 21),
+    get_signals_all(dat, "cusum", stratification = "age_group", number_of_time_units = 21),
     "The stratum age_group:15-19"
   )
 })